map view and main list merge
